+++ src/string/encoding.rs
@@ -36,7 +36,7 @@ impl PartialEq<AnyObject> for Encoding {
 unsafe impl Object for Encoding {
     #[inline]
     fn cast<A: Object>(obj: A) -> Option<Self> {
-        if obj.class().inherits(Class::encoding()) {
+        if obj.class().inherits(Class::of::<Self>()) {
             unsafe { Some(Self::cast_unchecked(obj)) }
         } else {
             None